+++ src/ga/base/test/RuleTests.java
@@ -8,7 +8,7 @@ public class RuleTests {
 
 	@Test
 	public void NumberLutConversion() throws Exception {
-		long i = (long)Integer.MAX_VALUE*2L+1;
+		long i = Integer.MAX_VALUE*2L+1;
 		Rule r = new Rule(i, 2);
 		if (r.getNumber() != i)
 			throw new Exception("fuck: " + i);
